Adding more detailed error messages.
Lexical error messages will include the filename of the file that the
invalid token was included as well as the line that the token is located
in the said file. This handling needed more thorough parsing of
multiline comments. I added a new parsing rule for multilined comments
called "multi_comments" that is heavily inspired by the last slide of
this set of slides: https://courses.softlab.ntua.gr/compilers/2019a/slides/slides-flex.pdf