First version of a "filename tracking" set. (Lexer)
We initialize the StringSet module as we've done before. Afterwards we
define a record type called globalset. It contains one mutable field
that is of type StringSet.t. This record will act as a wrapper and will
enable us to have a global "filename tracking" set. We initialize a
record of the said type (globalset). We also define in the header section
a function called "safe_find" because Set.find returns a Non_Found exception
if it fails to find the value we are searching for. We use this
safe_find to check if the latest filename the user tries to include has
been included in the past. If it has, we ignore that inclusion.
This strategy allows us to deal with cyclical inclusion between files
and also unnecessary multiple inclusion of one file.