A fix about the #include directive.
After trying many things that would keep the lexical analysis even if
the user wrote an incorrect #include directive, I concluded to the
following "solution". When the lexical analyser "munches" an incl
instance, we check via the position record of our Lexing.lexbuf if we
the incl munch started at the beginning of the line or not. If yes,
continue executing normally as before, otherwise show a proper error
message and afterwards terminate the lexical analysis.