A first attempt at an executable parser
Issues:
- Not good error handling in main.ml
- Should we do lexical analysis and syntactic analysis in parallel?
  Or should we not proceed into syntactic analysis in case of a lexical
  error?
- Not convinved about the solutions I've given for function calls, array
  access, type casting, the ternary operator etc. in the syntactic
  analyser.
- There is a reduce/reduce conflict in the grammar to begin with:

  expression_list := expression (, expression)*

  VS

  expression := expression , expression

  I am not quite confident in the quick and dirty solution I've given to
  this conflict.
- I cannot say I am a big fan of the "Lexer.init" function I've wrote
  and use in "Main.ml" and I would like to re-examine that if we have
  the luxury of doing so.

(Note: I've tested the current parser version with ~800+ programs from
the generator that we have been provided and at least it successfully
accepts those. I've tested the parser with not many syntactically
incorrect programs and it also seems to identify those. I am not 100%
sure of course about the correctness of our parser, so I would
appreciate if we could test the parser more and discuss solutions to the
problems that I stated above.)