Changes to the lexer
Abandoned the idea of using a stack. Now we depend on OCaml's recursion
when including a file. Also introduced the notion of a T_include "token"
as a placeholder for the AST that will be produced when a file is
included. Also abandoned the complex stack interface because it was
highly complicated and did not work properly. I've added some printing
messages so that I was able to see what tokens the parser would consume
via the lexer. I propose that we keep them for the time being and delete
them when necessary.