Declaration semantic analysis (Nothing compilable)
Started implementing the semantic analysis for the declarations. Have
not checked if Semantic.ml is compilable or not. Have to do some
exception handling and complete the analysis of function definitions.
Afterwards (if everything went "well"), I have to implement the semantic
analysis for the for loop, the break & continue statements and last but
not least the return statement.