Important changes in semantic analysis
Fixed the messages of "add_parameters". Introduced the function
"ent_name_of_id". This function is not needed, but I did not want to
flood the function "add_parameters" with one more parameter for the
function name. I created an .mli for SemUtilities.
And last but more important, I tried to enforce that when we declare
static arrays, the "size" of the array is actualy a constant integer
expression.